test(cli): move the scope path tests into the existing asgi test dir
Add the scope-echo app and its tests to the asgi directory instead of a
new one, so the HTTP ASGI tests share one workerd invocation.
