--- packages/cli/tests/workerd-test/asgi-scope/asgi-scope.wd-test
@@ -1,19 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "python-asgi-scope",
-      worker = (
-        modules = [
-          (name = "worker.py", pythonModule = embed "worker.py"),
-          %PYTHON_MODULES
-        ],
-        bindings = [
-          ( name = "SELF", service = "python-asgi-scope" ),
-        ],
-        compatibilityDate = "%COMPAT_DATE",
-        compatibilityFlags = ["python_workers"],
-      )
-    )
-  ],
-);
