--- packages/cli/tests/workerd-test/asgi-scope/pyproject.toml
@@ -1,5 +1,0 @@
-[project]
-name = "test"
-version = "0.0.0"
-requires-python = ">=3.12"
-dependencies = ["pytest", "pytest-asyncio<1.2.0"]
