--- packages/cli/tests/workerd-test/asgi-scope/wrangler.jsonc
@@ -1,5 +1,0 @@
-{
-    "name": "test-worker",
-    "compatibility_date": "%COMPAT_DATE",
-    "compatibility_flags": ["python_workers"]
-}
